requirements.txt
jieba==0.42.1
joblib==1.4.2
numpy==2.0.0
opencv_python==4.9.0.80
paddleocr==2.7.3
Pillow==9.3.0
Pillow==10.3.0
scikit_learn==1.4.2
skimage==0.0
torch==2.4.0.dev20240411
tqdm==4.66.2
transformers==4.41.2
zhon==2.0.2
